Changed size of the movement timer so it is onlt 16 bits, to save memory. Not that I'm anywhere near out of memory yet, but still.

+++ Games/TinyShooter/src/Ingame/Enemy.h
@@ -36,7 +36,7 @@ public:
 
 	fp2d m_worldPosition;
 	fp2d m_movementDirection;
-	int m_movementTimer;
+	uint16 m_movementTimer;
 
 	Enemy();
 	void Reboot();
